Starting to refactor the powerups
