+++ src/entities/PowerupBomb.hpp
@@ -24,6 +24,7 @@ THE SOFTWARE.
 
 #include "Powerup.hpp"
 #include "WeaponBomb.hpp"
+#include "components/Powerup.hpp"
 #include "misc/math.hpp"
 #include "services/ConfigurationPath.hpp"
 
@@ -33,11 +34,10 @@ namespace entities
     {
       public:
         PowerupBomb(math::Point2f position) :
-            Powerup(config::ENTITY_WEAPON_BOMB, position)
+            Powerup(components::Powerup::Type::Bomb, config::ENTITY_WEAPON_BOMB, position)
         {
         }
 
         virtual std::shared_ptr<Weapon> get() override { return std::make_shared<entities::WeaponBomb>(config::ENTITY_WEAPON_BOMB); }
-        virtual Type getType() override { return Type::Bomb; }
     };
 } // namespace entities
